feat: make test a unit test, add ignore_unknown_field
